+++ Classes/Triangle.cs
@@ -9,7 +9,7 @@ namespace Week1ObjectOriented.Classes
 {
     public class Triangle : Shape, IFaces
     {
-        public override int GetArea()
+        public override double GetArea()
         {
             throw new NotImplementedException();
         }
